make lotto num 클래스 분리
